+++ src project/Pages/Component/CollectionManager/ShowCollectionCard.razor
@@ -1,7 +1,10 @@
 ﻿@code {
     MongoDB_Web.Pages.Component.CollectionManager.DeleteCollectionModal? deleteCollectionModal { get; set; }
+    MongoDB_Web.Pages.Component.Util.DownloadModal? downloadModal { get; set; }
 }
-<DeleteCollectionModal @ref="deleteCollectionModal" DBName="@DBName" CollectionName="@name" />
+<DeleteCollectionModal @ref="deleteCollectionModal" dbName="@dbName" collectionName="@name" />
+<MongoDB_Web.Pages.Component.Util.DownloadModal @ref="downloadModal" dbName="@dbName"  collectionName="@name"/>
+
 <div class="col-lg-4 col-xl-3 col-md-5 mb-8 col-sm-12 py-3">
     <div class="shadow card h-100 rounded cardbox">
         <div class="card-header py-1">
@@ -23,12 +26,12 @@
                 <ul class="dropdown-menu element-background">
                     <li><a class="dropdown-item element-background" href="#">Details</a></li>
                     <li><a class="dropdown-item element-background" href="#">Upload .JSON</a></li>
-                    <li><a class="dropdown-item element-background" href="/Download/@DBName/@name/">Download Collection</a></li>
+                    <li><a class="dropdown-item element-background" href="" @onclick="() => downloadModal?.Open()" @onclick:preventDefault>Download Collection</a></li>
                     <li><hr class="dropdown-divider"></li>
                     <li><a class="dropdown-item text-danger" href="" @onclick="() => deleteCollectionModal?.Open()" @onclick:preventDefault>Delete Collection</a></li>
                 </ul>
             </div>
-            <a href="/CollectionViewer/@DBName/@name" class="btn btn-primary">Open</a>
+            <a href="/CollectionViewer/@dbName/@name" class="btn btn-primary">Open</a>
         </div>
     </div>
 </div>
@@ -36,7 +39,7 @@
     [Parameter]
     public MongoDB.Bson.BsonDocument CollectionObject { get; set; } = default!;
     [Parameter]
-    public string DBName { get; set; } = "";
+    public string dbName { get; set; } = "";
 
     string? name { get; set; } = "";
     bool? readOnly { get; set; } = null;
