+++ src project/wwwroot/js/action.js
@@ -30,42 +30,3 @@ function downloadURI (uri, name) {
     link.click();
     document.body.removeChild(link);
 }
-
-async function startConnection() {
-    const connection = new signalR.HubConnectionBuilder().withUrl("/progressHub").build();
-
-    connection.on("ReceiveProgressDatabase", function (totalCollections, processedCollections, progress) {
-        const progressBar = document.getElementById("fileProgress");
-        if (progressBar) {
-            progressBar.value = progress;
-        }
-
-        const progressText = document.getElementById("status-text");
-        if (progressText) {
-            progressText.innerHTML = processedCollections + " / " + totalCollections + " collections";
-        }
-    });
-
-    connection.on("ReceiveProgressCollection", function (totalDocuments, processedDocuments, progress) {
-        const progressBar = document.getElementById("fileProgress");
-        if (progressBar) {
-            progressBar.value = progress;
-        }
-
-        const progressText = document.getElementById("status-text");
-        if (progressText) {
-            progressText.innerHTML = processedDocuments + " / " + totalDocuments + " documents";
-        }
-    });
-
-    try {
-        await connection.start();
-        console.log("SignalR Connected.");
-    } catch (err) {
-        console.log(err);
-        setTimeout(startConnection, 5000);
-    }
-}
-
-startConnection();
-
